+++ src/main/java/com/example/BlogPost/repository/PostJPARepository.java
@@ -28,7 +28,7 @@ public class PostJPARepository implements PostRepository{
     }
 
     @Override
-    public Optional<Post> findById(Integer id) {
+    public Optional<Post> findById(Long id) {
         Post post = em.find(Post.class, id);
         return Optional.ofNullable(post);
     }
@@ -58,7 +58,7 @@ public class PostJPARepository implements PostRepository{
     }
 
     @Override
-    public Integer deleteById(Integer id) {
+    public Integer deleteById(Long id) {
         Query query = em.createQuery("DELETE FROM Post p WHERE p.post_id = :id");
         query.setParameter("id", id);
