+++ src/main/java/com/example/BlogPost/repository/PostRepository.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 
 public interface PostRepository {
     Post upload(PostDTO post);
-    Optional<Post> findById(Integer id);
+    Optional<Post> findById(Long id);
     Optional<Post> findByTitle(String title);
     List<Post> findAll();
 
     Post edit(Post post);
-    Integer deleteById(Integer id);
+    Integer deleteById(Long id);
 }
